Additional Commit to Update Some Alterations (6.26)
1. Decided to change button layout for HUD. 8 rects just takes up way
too much space, so I figured we can use arrows in spaces between the
up/down/right/left buttons. The arrows will be for diagonal movement and
will be in a wrapped bitmap class with rects to track location for user
touch. I want to keep the array of rect's and array of wrapped bitmap
objects in the same object, so we could possibly do that with a pair
object (one array is key, the other is value). Game compiled and ran
properly before changes however.

This will probably be the biggest obstacle in getting this done other
than implementing x/y-clip for background movement both ways. Overall,
feel like we are in pretty good place though.